+++ libs/cardano-ledger-core/src/Cardano/Ledger/PoolDistr.hs
@@ -4,11 +4,9 @@
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
-{-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE UndecidableInstances #-}
 
 -- | The stake distribution, aggregated by stake pool (as opposed to stake credential),
@@ -23,7 +21,7 @@ module Cardano.Ledger.PoolDistr (
   PoolDistr (..),
   poolDistrDistrL,
   poolDistrTotalL,
-  individualPoolStakeCoinL,
+  individualTotalPoolStakeL,
 )
 where
 
@@ -54,19 +52,25 @@ import NoThunks.Class (NoThunks (..))
 -- delegated to a registered stake pool.
 data IndividualPoolStake c = IndividualPoolStake
   { individualPoolStake :: !Rational
-  , individualPoolStakeCoin :: !(CompactForm Coin) -- TODO: Document
+  -- ^ Pool stake distribution. This is a ratio of
+  -- (individualTotalPoolStake - proposal deposits) to
+  -- (pdTotalActiveStake - proposal deposits)
+  , individualTotalPoolStake :: !(CompactForm Coin)
+  -- ^ Total stake delegated to this pool. In addition to all the stake  that
+  -- is part of `individualPoolStake` we also add proposal-deposits to this
+  -- field.
   , individualPoolStakeVrf :: !(Hash c (VerKeyVRF c))
   }
   deriving stock (Show, Eq, Generic)
   deriving anyclass (NFData, NoThunks)
 
-individualPoolStakeCoinL :: Lens' (IndividualPoolStake c) (CompactForm Coin)
-individualPoolStakeCoinL = lens individualPoolStakeCoin $ \x y -> x {individualPoolStakeCoin = y}
+individualTotalPoolStakeL :: Lens' (IndividualPoolStake c) (CompactForm Coin)
+individualTotalPoolStakeL = lens individualTotalPoolStake $ \x y -> x {individualTotalPoolStake = y}
 
 instance Crypto c => EncCBOR (IndividualPoolStake c) where
   encCBOR (IndividualPoolStake stake stakeCoin vrf) =
     mconcat
-      [ encodeListLen 3 -- TODO: This needs versioning!
+      [ encodeListLen 3
       , encCBOR stake
       , encCBOR stakeCoin
       , encCBOR vrf
@@ -74,7 +78,7 @@ instance Crypto c => EncCBOR (IndividualPoolStake c) where
 
 instance Crypto c => DecCBOR (IndividualPoolStake c) where
   decCBOR =
-    decodeRecordNamed "IndividualPoolStake" (const 2) $
+    decodeRecordNamed "IndividualPoolStake" (const 3) $
       IndividualPoolStake
         <$> decCBOR
         <*> decCBOR
@@ -88,15 +92,19 @@ toIndividualPoolStakePair :: (KeyValue e a, Crypto c) => IndividualPoolStake c -
 toIndividualPoolStakePair indivPoolStake@(IndividualPoolStake _ _ _) =
   let IndividualPoolStake {..} = indivPoolStake
    in [ "individualPoolStake" .= individualPoolStake
-      , "individualPoolStakeCoin" .= individualPoolStakeCoin
+      , "individualTotalPoolStake" .= individualTotalPoolStake
       , "individualPoolStakeVrf" .= individualPoolStakeVrf
       ]
 
 -- | A map of stake pool IDs (the hash of the stake pool operator's
--- verification key) to 'IndividualPoolStake'.
+-- verification key) to 'IndividualPoolStake'. Also holds absolute values
+-- necessary for the calculations in the `computeDRepDistr`.
 data PoolDistr c = PoolDistr
-  { unPoolDistr :: Map (KeyHash 'StakePool c) (IndividualPoolStake c)
-  , pdTotal :: !(CompactForm Coin)
+  { unPoolDistr :: !(Map (KeyHash 'StakePool c) (IndividualPoolStake c))
+  , pdTotalActiveStake :: !(CompactForm Coin)
+  -- ^ Total stake delegated to registered stake pools. In addition to
+  -- the stake considered for the `individualPoolStake` Rational, we add
+  -- proposal-deposits to this field.
   }
   deriving stock (Show, Eq, Generic)
   deriving (NFData, NoThunks, ToJSON)
@@ -105,11 +113,10 @@ poolDistrDistrL :: Lens' (PoolDistr c) (Map (KeyHash 'StakePool c) (IndividualPo
 poolDistrDistrL = lens unPoolDistr $ \x y -> x {unPoolDistr = y}
 
 poolDistrTotalL :: Lens' (PoolDistr c) (CompactForm Coin)
-poolDistrTotalL = lens pdTotal $ \x y -> x {pdTotal = y}
+poolDistrTotalL = lens pdTotalActiveStake $ \x y -> x {pdTotalActiveStake = y}
 
 instance Crypto c => EncCBOR (PoolDistr c) where
   encCBOR (PoolDistr distr total) =
-    -- TODO: @aniketd: This needs versioning!
     encode $
       Rec PoolDistr
         !> To distr
@@ -121,27 +128,3 @@ instance Crypto c => DecCBOR (PoolDistr c) where
       RecD PoolDistr
         <! From
         <! From
-
--- -- ===============================
-
--- instance
---   HasExp
---     (PoolDistr c)
---     ( Map
---         (KeyHash 'StakePool c)
---         (IndividualPoolStake c)
---     )
---   where
---   toExp (PoolDistr x) = Base MapR x
-
--- -- | We can Embed a Newtype around a Map (or other Iterable type) and then use it in a set expression.
--- instance
---   Embed
---     (PoolDistr c)
---     ( Map
---         (KeyHash 'StakePool c)
---         (IndividualPoolStake c)
---     )
---   where
---   toBase (PoolDistr x) = x
---   fromBase = PoolDistr