+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Classes.hs
@@ -365,7 +365,12 @@ class (Show x, Adds x) => Sums t x | t -> x where
 
 instance GoodCrypto c => Sums (IndividualPoolStake c) Rational where
   getSum (IndividualPoolStake r _ _) = r
-  genT _ r = IndividualPoolStake r mempty <$> arbitrary
+  genT _ r =
+    -- We use mempty for the individualTotalPoolStake here
+    -- but that field is intended to hold the total stake
+    -- assigned to the pool. We do not have enough information
+    -- here to be able to assign it its correct value.
+    IndividualPoolStake r mempty <$> arbitrary
 
 instance Reflect era => Sums (TxOutF era) Coin where
   getSum (TxOutF _ txout) = coin (txout ^. valueTxOutL)